requirements.txt
bcrypt==4.2.0
blinker==1.8.2
click==8.1.7
flask==3.0.3
Flask-Cors==5.0.0
Flask-Login==0.6.3
flask-wtf==1.2.1
greenlet==3.0.3
gunicorn==23.0.0
importlib-metadata==8.4.0
itsdangerous==2.2.0
jinja2==3.1.4
MarkupSafe==2.1.5
mysql-connector==2.2.9
mysql-connector-python==9.0.0
mysqlclient==2.2.4
packaging==24.1
pillow==10.4.0
PyMySQL==1.1.1
python-dotenv==1.0.1
SQLAlchemy==2.0.32
typing-extensions==4.12.2
werkzeug==3.0.4
wtforms==3.1.2
zipp==3.20.1
